Feature #18 : 서버 통신에 사용될 data class인 CGVTimeTable과 Theaters 생성
