+++ app/src/main/java/org/sopt/cgv/feature/seats/data/Theaters.kt
@@ -1,0 +1,6 @@
+package org.sopt.cgv.feature.seats.data
+
+data class Theaters(
+    val id: Int,
+    val theaterName: String
+)
